refactor: common wrapper for tab lists
